move populate_signature_symbols to trait method
